Change decorators.py to utils.py and add a "is_prime" function to utils.py
